# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LeagueLift API
  description: >
    LeagueLift revenue, fundraising, commerce, and fee-management API for youth sports
    organizations. This document covers Phase 0 (foundation), Phase 1 (organization
    onboarding, teams, tournaments, and public pages), Phase 2 (households,
    participants, fee templates, and fee assignments), Phase 3 slice 11
    (fundraising campaigns and their public pages) plus its remainder (campaign
    contribution recording via test-mode Stripe Checkout, real raised totals, and
    an inbound Stripe webhook receiver — DESIGN-DOC.md section 8.5; family
    credit rules remain a later slice pending section 19.3's open questions),
    traditional email/password authentication (superseding
    the originally planned external-IdP/Auth0 approach — see DESIGN-DOC.md section
    11.5), Phase 1's remaining file-upload/branding slice (organization logo/cover
    image via S3-compatible object storage — DESIGN-DOC.md section 11, ADR-012), and
    Phase 2's remainder (manual/offline fee payments, manual discounts/credits,
    org-wide collections listing and CSV export, and Stripe Connect Express
    onboarding-only scaffolding — DESIGN-DOC.md section 14.3, ADR-005 — no live
    charge routing yet), and Phase 4 slice 1 (apparel commerce proof: stores,
    Printify-catalog-backed products/variants with real Printify cost snapshots,
    a Printify catalog read proxy, multi-item test-mode Stripe Checkout with US
...
servers:
...
paths:
...
  /organizations/{organizationId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationIdParam"
    get:
      tags: [organizations]
      summary: Get an organization by ID
      operationId: getOrganization
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [organizations]
      summary: Update an organization
      operationId: updateOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/members:
    parameters:
      - $ref: "#/components/parameters/OrganizationIdParam"
    get:
      tags: [members]
      summary: List members of an organization
      operationId: listOrganizationMembers
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
      responses:
        "200":
          description: Paginated list of memberships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/members/{memberId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationIdParam"
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [members]
      summary: Change a member's role
      description: >
        Owner-only role is out of scope here — ownership transfer is a separate
        controlled workflow (DESIGN-DOC.md section 7.2). Only OWNER or ADMINISTRATOR
        callers may change other members' roles.
      operationId: updateMembershipRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMembershipRoleRequest"
      responses:
        "204":
          description: Role updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [members]
      summary: Revoke a member's access
      operationId: revokeMembership
      responses:
        "204":
          description: Membership revoked
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /organizations/{organizationId}/media/uploads:
    parameters:
      - $ref: "#/components/parameters/OrganizationIdParam"
    post:
      tags: [media]
      summary: Request a presigned upload URL
      description: >
        Creates a PENDING_UPLOAD media asset and returns a short-lived presigned PUT
        URL. The browser uploads directly to object storage; the API never proxies
        file bytes.
      operationId: requestMediaUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadRequest"
      responses:
        "201":
          description: Upload requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /organizations/{organizationId}/documents:
    parameters:
      - $ref: "#/components/parameters/OrganizationIdParam"
    get:
      tags: [documents]
      summary: List the organization's documents (any active org member)
      operationId: listOrganizationDocuments
      responses:
        "200":
          description: Organization documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [documents]
      summary: Assign an already-uploaded (usageSlot=DOCUMENT) asset as an organization-level document (org manager only)
      description: Upload the underlying file first via POST /organizations/{organizationId}/media/uploads (usageSlot=DOCUMENT) + its /confirm call.
      operationId: assignOrganizationDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDocumentRequest"
      responses:
        "201":
          description: Document assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/documents/broadcast:
    post:
      tags: [documents]
      summary: Assign an already-uploaded document to every active household in the organization (org manager only)
      description: Convenience for a season-wide waiver/form — creates one assignment per household, all sharing the same underlying asset.
      operationId: broadcastDocumentToAllHouseholds
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDocumentRequest"
      responses:
        "201":
          description: Document assigned to every household
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/documents/{assignmentId}:
    delete:
      tags: [documents]
      summary: Retire a document (org manager only) — never deletes the underlying asset, which may back other households' assignments from a broadcast
      operationId: removeDocument
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Retired
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/documents/{assignmentId}/acknowledgments:
    get:
      tags: [documents]
      summary: List which guardians have acknowledged/signed this document (org manager only)
      operationId: listDocumentAcknowledgments
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Acknowledgments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentAcknowledgmentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/households/{householdId}/documents:
    parameters:
      - $ref: "#/components/parameters/OrganizationIdParam"
      - $ref: "#/components/parameters/HouseholdIdParam"
    get:
      tags: [documents]
      summary: List a household's documents (org member or a guardian of this household)
      operationId: listHouseholdDocuments
      responses:
        "200":
          description: Household documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [documents]
      summary: Assign an already-uploaded document to one household (org manager only)
      operationId: assignHouseholdDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDocumentRequest"
      responses:
        "201":
          description: Document assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/households/{householdId}/documents/{assignmentId}/acknowledge:
    post:
      tags: [documents]
      summary: Acknowledge ("sign") a household document — must be an active guardian of this household. Idempotent — re-acknowledging returns the existing record.
      operationId: acknowledgeDocument
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/HouseholdIdParam"
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Acknowledgment recorded (or already existed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentAcknowledgmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /organizations/{organizationId}/search:
    get:
      tags: [search]
      summary: Global search within one organization — teams, participants, and households (DESIGN-DOC.md section 13, Phase 7 completion; any active org member)
      operationId: searchOrganization
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results (empty for a query under 2 characters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /platform/search:
    get:
      tags: [search]
      summary: Global search across organizations, platform-wide (platform administrator only)
      operationId: searchPlatform
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results (empty for a query under 2 characters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationIdParam:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    HouseholdIdParam:
      name: householdId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found (or not visible to the caller)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, requestId, fieldErrors]
      properties:
        code:
          type: string
          example: ORGANIZATION_NOT_FOUND
        message:
          type: string
          example: The organization could not be found.
        requestId:
          type: string
          example: req_01hz8x9y2z
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        organizationType:
          $ref: "#/components/schemas/OrganizationType"
        sports:
          type: array
          items:
            type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
          maxLength: 40
    OrganizationType:
      type: string
      enum:
        - RECREATIONAL_LEAGUE
        - TRAVEL_CLUB
        - INDIVIDUAL_TEAM
        - TOURNAMENT_OPERATOR
        - BOOSTER_ORGANIZATION
        - MULTISPORT_FACILITY
        - COMMUNITY_PROGRAM
        - OTHER
    OrganizationResponse:
      type: object
      required: [id, name, slug, organizationType, status, sports, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        organizationType:
          $ref: "#/components/schemas/OrganizationType"
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, ARCHIVED]
        sports:
          type: array
          items:
            type: string
        contactEmail:
          type: string
          format: email
          nullable: true
        contactPhone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MembershipRole:
      type: string
      enum:
        - OWNER
        - ADMINISTRATOR
        - FINANCE_MANAGER
        - TEAM_ADMINISTRATOR
        - TOURNAMENT_ADMINISTRATOR
        - VIEWER
    MembershipResponse:
      type: object
      required: [id, organizationId, userId, role, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        userEmail:
          type: string
          nullable: true
        userDisplayName:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/MembershipRole"
        status:
          type: string
          enum: [ACTIVE, INVITED, REVOKED]
        createdAt:
          type: string
          format: date-time
    MembershipPage:
      type: object
      required: [items, page, size, totalElements]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MembershipResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
    UpdateMembershipRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/MembershipRole"
...
    RequestUploadRequest:
      type: object
      required: [usageSlot, fileName, contentType, fileSizeBytes]
      properties:
        usageSlot:
          type: string
          enum: [LOGO, COVER]
        fileName:
          type: string
        contentType:
          type: string
          example: image/png
        fileSizeBytes:
          type: integer
          format: int64
    RequestUploadResponse:
      type: object
      required: [assetId, uploadUrl, uploadMethod, requiredHeaders, expiresAt]
      properties:
        assetId:
          type: string
          format: uuid
        uploadUrl:
          type: string
        uploadMethod:
          type: string
          example: PUT
        requiredHeaders:
          type: object
          additionalProperties:
            type: string
        expiresAt:
          type: string
          format: date-time
...
    AssignDocumentRequest:
      type: object
      required: [assetId]
      properties:
        assetId:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
    DocumentResponse:
      type: object
      required: [id, assetId, url, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        assetId:
          type: string
          format: uuid
        url:
          type: string
        title:
          type: string
          nullable: true
        contentType:
          type: string
          nullable: true
        byteSizeBytes:
          type: integer
          format: int64
          nullable: true
        updatedAt:
          type: string
          format: date-time
    DocumentListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentResponse"
    DocumentAcknowledgmentResponse:
      type: object
      required: [id, householdAdultId, acknowledgedByUserId, acknowledgedAt]
      properties:
        id:
          type: string
          format: uuid
        householdAdultId:
          type: string
          format: uuid
        acknowledgedByUserId:
          type: string
          format: uuid
        acknowledgedAt:
          type: string
          format: date-time
    DocumentAcknowledgmentListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentAcknowledgmentResponse"

    # --- Payouts ---
...
    SearchHitResponse:
      type: object
      required: [type, id, label]
      properties:
        type:
          type: string
          enum: [TEAM, PARTICIPANT, HOUSEHOLD, ORGANIZATION]
        id:
          type: string
          format: uuid
        label:
          type: string
        subtitle:
          type: string
          nullable: true
    SearchResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SearchHitResponse"
...
security:
  - bearerAuth: []